feat: add option to clone env params from src env (move to env params package)
