+++ packages/amplify-provider-awscloudformation/src/index.ts
@@ -51,9 +51,6 @@ import { hashDirectory } from './upload-appsync-files';
 import { prePushCfnTemplateModifier } from './pre-push-cfn-processor/pre-push-cfn-modifier';
 import { getApiKeyConfig } from './utils/api-key-helpers';
 
-import { cloneEnvParamManager } from './utils/clone-env-param-manager';
-export { cloneEnvParamManager } from './utils/clone-env-param-manager';
-
 function init(context) {
   return initializer.run(context);
 }
@@ -201,5 +198,4 @@ module.exports = {
   hashDirectory,
   prePushCfnTemplateModifier,
   getApiKeyConfig,
-  cloneEnvParamManager,
 };
